+++ yojimbo_client.h
@@ -205,8 +205,6 @@ namespace yojimbo
 
         virtual MessageFactory * CreateMessageFactory( Allocator & allocator );
 
-        virtual void InitializeContext( const ConnectionConfig & connectionConfig, MessageFactory & messageFactory );
-
         virtual void SetClientState( int clientState );
 
         virtual void ResetConnectionData( int clientState = CLIENT_STATE_DISCONNECTED );
@@ -277,7 +275,9 @@ namespace yojimbo
 
         int m_clientIndex;                                                  // the client index on the server [0,maxClients-1]. -1 if not connected.
 
-        ClientServerContext m_context;                                      // serialization context for client/server packets.
+        TransportContext m_transportContext;                                // transport context for reading and writing packets.
+
+        ConnectionContext m_connectionContext;                              // connection context for serializing connection packets.
 
         ClientState m_clientState;                                          // current client state